fleet_master/requirements.txt
fastapi==0.115.12                 # 框架核心
uvicorn==0.34.0                   # ASGI server，用于启动 FastAPI
sqlmodel==0.0.24                  # ORM 框架，Pydantic + SQLAlchemy 封装
SQLAlchemy==2.0.40                # 被 sqlmodel 底层依赖
pydantic==2.11.2                  # 数据模型校验
pydantic_settings==2.8.1          # 配置管理（Pydantic v2 的配置方式）
alembic==1.15.2                   # 数据库迁移工具（如果你用）
python-jose==3.4.0                # JWT 库
asyncpg==0.30.0                   # 异步 PostgreSQL 客户端
python-multipart==0.0.6           # 处理表单
greenlet==3.1.1                   # 协程库
bcrypt==4.3.0                     # 密码哈希
httpx==0.28.1                     # HTTP 客户端
orjson==3.10.16                   # JSON 序列化/反序列化
gunicorn==23.0.0                  # 生产环境部署
psutil==7.0.0                     # 系统监控
apscheduler==3.11.0               # 任务调度
tzlocal==5.3.1                    # 时区处理